music_service/requirements.txt
flask==2.3.3
kafka-python==2.0.2
python-dotenv==1.0.0
flask-cors==4.0.0